refactor: :truck: appディレクトリ内を整理
