+++ packages/app/src/app/main.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { App } from "./App";
+import { App } from "./ui/App";
 
 const root = document.getElementById("root");
 if (root) {
