+++ packages/app/src/app/ui/App.tsx
@@ -1,5 +1,5 @@
-import "./global.css";
-import { SceneSwitcher } from "../components/scenes/SceneSwitcher";
+import "../styles/global.css";
+import { SceneSwitcher } from "./SceneSwitcher";
 import "@/shared/i18n/config";
 import i18next from "i18next";
 import type { FC } from "react";
