+++ packages/app/src/app/ui/SceneSwitcher.tsx
@@ -1,7 +1,7 @@
+import { GameScene } from "@/components/scenes/GameScene";
+import { ModeSelectScene } from "@/components/scenes/ModeSelectScene";
 import { useScene } from "@/game/hooks/useScene";
 import type { FC } from "react";
-import { GameScene } from "./GameScene";
-import { ModeSelectScene } from "./ModeSelectScene";
 
 export const SceneSwitcher: FC = () => {
   const { scene } = useScene();
